requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
jsonschema>=4.23.0
PyYAML>=6.0.2
tabulate>=0.9.0
tenacity>=9.0.0
rich>=13.8.1

